Minor Additions to TowerScript
Addded a line of code to set the sphere collider to the range in
Start(). Added some if-statements to check if the collider is tagged as
an "Enemy" in OnTriggerEnter() and OnTriggerExit(). Fixed an issue where
the list was not initialized, causing a null reference exception to be
thrown every frame.