A lot was added.
The wall is now done. Towers now have colors to help distinguish them
more (I even got a trasnparent effect for the energy part working). Wall
Nodes and normal Nodes are now copmlete, and set up around a mock level
that we can use for the prototype. A sniper nest was made for the camera
to be placed. You can now build all 3 base towers in tower build mode
and functions properly with all the node types.